@@ -1,28 +1,27 @@
 {
   "name": "metadata-portal",
   "version": "0.1.0",
-  "homepage": "https://centrifuge.github.io/metadata-portal",
   "private": true,
   "dependencies": {
-    "@headlessui/react": "^1.4.2",
-    "@heroicons/react": "^1.0.5",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.0.0",
-    "@testing-library/user-event": "^13.2.1",
-    "@types/jest": "^27.0.1",
-    "@types/node": "^16.7.13",
-    "@types/react": "^17.0.20",
-    "@types/react-dom": "^17.0.9",
-    "husky": "^7.0.4",
-    "lint-staged": "^12.1.4",
-    "prettier": "^2.5.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-github-fork-ribbon": "^0.6.0",
-    "react-router-dom": "^6.2.1",
-    "react-scripts": "5.0.0",
-    "typescript": "^4.5.5",
-    "web-vitals": "^2.1.0"
+    "@headlessui/react": "^1.7.10",
+    "@heroicons/react": "^2.0.14",
+    "@polkadot/extension-dapp": "^0.44.8",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^14.4.3",
+    "@types/jest": "^29.4.0",
+    "@types/node": "^18.13.0",
+    "@types/react": "^18.0.27",
+    "@types/react-dom": "^18.0.10",
+    "husky": "^8.0.3",
+    "lint-staged": "^13.1.1",
+    "prettier": "^2.8.4",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.8.1",
+    "react-scripts": "5.0.1",
+    "typescript": "^4.9.5",
+    "web-vitals": "^3.1.1"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -30,6 +29,8 @@
     "predeploy": "npm run build",
     "deploy": "gh-pages -d build",
     "test": "react-scripts test",
+    "lint": "eslint --ext .ts,.tsx src",
+    "prettier": "prettier --write 'src/**/*.(tsx|ts)'",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
@@ -61,14 +62,14 @@
     ]
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^5.11.0",
-    "@typescript-eslint/parser": "^5.11.0",
-    "autoprefixer": "^10.4.0",
-    "eslint": "^8.8.0",
-    "eslint-plugin-react": "^7.28.0",
-    "gh-pages": "^3.2.3",
-    "postcss": "^8.4.5",
-    "prettier-eslint": "^13.0.0",
-    "tailwindcss": "^3.0.7"
+    "@typescript-eslint/eslint-plugin": "^6.10.0",
+    "@typescript-eslint/parser": "^6.10.0",
+    "autoprefixer": "^10.4.16",
+    "eslint": "^8.53.0",
+    "eslint-plugin-react": "^7.33.2",
+    "gh-pages": "^5.0.0",
+    "postcss": "^8.4.31",
+    "prettier-eslint": "^16.1.2",
+    "tailwindcss": "^3.3.5"
   }
 }
