@@ -1,9 +1,15 @@
-import React from "react";
-import { render, screen } from "@testing-library/react";
+import { render } from "@testing-library/react";
 import App from "./App";
+import { BrowserRouter as Router } from "react-router-dom";
 
-test("renders learn react link", () => {
-  render(<App />);
-  const linkElement = screen.getByText(/learn react/i);
-  expect(linkElement).toBeInTheDocument();
+test("renders ok", () => {
+  render(
+    <Router>
+      <App />
+    </Router>
+  );
+});
+
+test("data file exists", async () => {
+  require("../../public/data.json");
 });
