@@ -1,35 +1,60 @@
-import jsonData from "./chains.json";  // Dynamically generated datafile. Run `make collector` to create
-
 export interface ChainSpec {
-  name: string;
-  iconPath: string;
+  title: string;
+  color: string;
   rpcEndpoint: string;
   genesisHash: string;
-  color: string;
   unit: string;
-  addressPrefix: number;
-  metadataQr?: QrInfo;
-  specsQr?: QrInfo;
+  base58prefix: number;
+  decimals: number;
+  logo: string;
+  liveMetaVersion: number;
+  metadataQr?: MetadataQr;
+  latestMetadata: string;
+  specsQr: QrInfo;
+  relayChain?: string;
 }
+
+export type SourceType = WasmSource | RpcSource | null;
 
 export interface QrInfo {
   path: string;
-  signedBy?: string;
-  version?: number;
+  signedBy: string | null;
+  source: SourceType;
+}
+
+export interface MetadataQr {
+  version: number;
+  file: QrInfo;
+  status: string;
+}
+
+interface SourceBase {
+  type: string;
+}
+
+export interface WasmSource extends SourceBase {
+  github_repo: string;
+  hash: string;
+}
+
+export interface RpcSource extends SourceBase {
+  block: string;
+}
+
+export interface AddToSignerInterface {
+  path: string;
+  color: string;
+  name: string;
 }
 
 export interface Chains {
   [name: string]: ChainSpec;
 }
 
-export function getChains(): Chains {
-  const chainList = jsonData.map((chain: object) =>
-    Object.assign({} as ChainSpec, chain)
-  );
-  return chainList.reduce((obj: Chains, chain: ChainSpec) => {
-    return {
-      ...obj,
-      [chain.name]: chain,
-    };
-  }, {});
+export type Portal = {
+  name: string;
+  url: string;
+};
+export interface Portals {
+  [name: string]: Portal;
 }
