@@ -1,7 +1,12 @@
 module.exports = {
   content: ["./src/**/*.{js,jsx,ts,tsx}"],
   theme: {
-    extend: {},
+    extend: {
+      borderRadius: {
+        "4xl": "2rem",
+      },
+    },
   },
   plugins: [],
+  important: true,
 };
